Move `Capability` to `irctext::types`
